Refactors Result
