+++ client/src/components/Result/index.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+import './Result.css';
+
+interface ResultProps {
+  shortDef: string[];
+}
+
+const Result = ({ shortDef }: ResultProps): JSX.Element => {
+  return (
+    <div className='result-container'>
+      <div className='result-header'>Definition</div>
+      {shortDef.map((def, i) => {
+        return (
+          <li className='definition' key={'def' + i}>
+            {def}
+          </li>
+        );
+      })}
+    </div>
+  );
+};
+
+export default Result;
